requirements.txt
yt-dlp
gallery-dl
pycryptodome
